requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
scikit-learn==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
requests==2.32.5
joblib==1.3.2
